+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Add a "TBD" component status and make it the default
+- Add component status to all component declaration templates
 
 
 ## [0.4.0] - 2018-03-16