+++ CHANGELOG.md
@@ -4,13 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.4.1] - unreleased
+## [0.5.0] - unreleased
 
 ### Added
 
 - Add a "TBD" component status and make it the default
 - Add component status to all component declaration templates
 
+### Changed
+
+- Renamed the default preview template
 
 ## [0.4.0] - 2018-03-16